Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    int a;
-    scanf("%d" , &a);
+    float a;
+    scanf("%f" , &a);
     float area = (float) (3.14*a*a);
     printf("Area: %.2f", area);
     return 0;
